dogshark/requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
fastapi==0.68.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.20.0
hyperframe==6.0.1
idna==3.3
loguru==0.5.3
nonebot-adapter-cqhttp==2.0.0a16
nonebot2==2.0.0a16
numpy==1.21.4
opencv-python-headless==4.5.3.56
Pillow==8.4.0
pydantic==1.8.2
pygtrie==2.4.2
python-dotenv==0.19.2
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.14.2
tomlkit==0.7.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
win32-setctime==1.0.3
